优化部分单元测试代码，以规避潜在的 bug
com.dwarfeng.notify.impl.service.DispatcherInfoMaintainServiceImplTest
com.dwarfeng.notify.impl.service.DispatcherSupportMaintainServiceImplTest
com.dwarfeng.notify.impl.service.MetaIndicatorMaintainServiceImplTest
com.dwarfeng.notify.impl.service.MetaMaintainServiceImplTest
com.dwarfeng.notify.impl.service.NotifyHistoryMaintainServiceImplTest
com.dwarfeng.notify.impl.service.NotifyInfoRecordMaintainServiceImplTest
com.dwarfeng.notify.impl.service.NotifySendRecordMaintainServiceImplTest
com.dwarfeng.notify.impl.service.NotifySettingMaintainServiceImplTest
com.dwarfeng.notify.impl.service.RouterInfoMaintainServiceImplTest
com.dwarfeng.notify.impl.service.RouterSupportMaintainServiceImplTest
com.dwarfeng.notify.impl.service.SenderInfoMaintainServiceImplTest
com.dwarfeng.notify.impl.service.SenderSupportMaintainServiceImplTest
com.dwarfeng.notify.impl.service.TopicMaintainServiceImplTest
com.dwarfeng.notify.impl.service.UserMaintainServiceImplTest

+++ notify-impl/src/test/java/com/dwarfeng/notify/impl/service/DispatcherInfoMaintainServiceImplTest.java
@@ -14,6 +14,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
+import java.util.Objects;
+
 import static org.junit.Assert.*;
 
 @RunWith(SpringJUnit4ClassRunner.class)
@@ -50,8 +52,12 @@ public class DispatcherInfoMaintainServiceImplTest {
             DispatcherInfo testDispatcherInfo = dispatcherInfoMaintainService.get(dispatcherInfo.getKey());
             assertEquals(BeanUtils.describe(dispatcherInfo), BeanUtils.describe(testDispatcherInfo));
         } finally {
-            dispatcherInfoMaintainService.deleteIfExists(dispatcherInfo.getKey());
-            topicMaintainService.deleteIfExists(topic.getKey());
+            if (Objects.nonNull(dispatcherInfo.getKey())) {
+                dispatcherInfoMaintainService.deleteIfExists(dispatcherInfo.getKey());
+            }
+            if (Objects.nonNull(topic.getKey())) {
+                topicMaintainService.deleteIfExists(topic.getKey());
+            }
         }
     }
 
@@ -68,8 +74,12 @@ public class DispatcherInfoMaintainServiceImplTest {
 
             assertFalse(dispatcherInfoMaintainService.exists(dispatcherInfo.getKey()));
         } finally {
-            dispatcherInfoMaintainService.deleteIfExists(dispatcherInfo.getKey());
-            topicMaintainService.deleteIfExists(topic.getKey());
+            if (Objects.nonNull(dispatcherInfo.getKey())) {
+                dispatcherInfoMaintainService.deleteIfExists(dispatcherInfo.getKey());
+            }
+            if (Objects.nonNull(topic.getKey())) {
+                topicMaintainService.deleteIfExists(topic.getKey());
+            }
         }
     }
 }
