+++ notify-impl/src/test/java/com/dwarfeng/notify/impl/service/DispatcherSupportMaintainServiceImplTest.java
@@ -14,6 +14,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 import static org.junit.Assert.assertEquals;
 
@@ -52,6 +53,9 @@ public class DispatcherSupportMaintainServiceImplTest {
             }
         } finally {
             for (DispatcherSupport dispatcherSupport : dispatcherSupports) {
+                if (Objects.isNull(dispatcherSupport.getKey())) {
+                    continue;
+                }
                 service.delete(dispatcherSupport.getKey());
             }
         }
