+++ notify-impl/src/test/java/com/dwarfeng/notify/impl/service/NotifySettingMaintainServiceImplTest.java
@@ -13,6 +13,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 import static org.junit.Assert.assertEquals;
 
@@ -50,6 +51,9 @@ public class NotifySettingMaintainServiceImplTest {
             }
         } finally {
             for (NotifySetting notifySetting : notifySettings) {
+                if (Objects.isNull(notifySetting.getKey())) {
+                    continue;
+                }
                 notifySettingMaintainService.deleteIfExists(notifySetting.getKey());
             }
         }
