+++ notify-impl/src/test/java/com/dwarfeng/notify/impl/service/TopicMaintainServiceImplTest.java
@@ -14,6 +14,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 import static org.junit.Assert.assertEquals;
 
@@ -50,6 +51,9 @@ public class TopicMaintainServiceImplTest {
             }
         } finally {
             for (Topic topic : topics) {
+                if (Objects.isNull(topic.getKey())) {
+                    continue;
+                }
                 service.delete(topic.getKey());
             }
         }
