+++ notify-impl/src/test/java/com/dwarfeng/notify/impl/service/UserMaintainServiceImplTest.java
@@ -14,6 +14,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 import static org.junit.Assert.assertEquals;
 
@@ -53,6 +54,9 @@ public class UserMaintainServiceImplTest {
             }
         } finally {
             for (User user : users) {
+                if (Objects.isNull(user.getKey())) {
+                    continue;
+                }
                 userMaintainService.deleteIfExists(user.getKey());
             }
         }
